Bug fix: selective where filters after join to time spine (#1120)
When an input metric specifies `join_to_timespine: true` in its YAML
definition, we apply special behavior for filters. Before this PR, we
were applying all filters twice - before and after the time spine join.
This is because you might end up with new rows after the time spine join
that should be filtered out. This resulted in a bug sometimes if you
filtered by a spec that was not in the group by, since the filter column
no longer existed at that level of the query.
This fixes that bug by only applying the post-join filter if the
filtered spec is also in the group by. The specs in the group by should
be the only ones changed after the time spine join, anyway.