Cache linkable elements (#1229)
Recently we've seen an increase in latency for compile time on queries.
The theory is that this might be caused by the addition of
`LinkableMetrics` to group by options, which may have dramatically
increased the number of group by options available. Looping through them
could add a lot of slowness to the query.

I added some logs to see how long those loops were taking, and looking
at data from the past 24 hours, the call to
`linkable_elements_for_measure` took about .1 seconds on each call for
long-running queries. This could be called quite a few times for the
same measure within a given query, resulting in a lot of duplicate work
& unnecessary latency. To fix that, I decided to cache that method. I
also cached a couple of similar methods along the way.

I tested this primarily by running queries in pytest and showing INFO
logs. There was a major reduction in the logs for these functions after
adding the cache (especially in scenarios where a given measure is used
several times in one query).

The other changes in this PR are making params hashable, which is
required for the cache decorator.
I did not add an expiration for the cache - since these classes are
scoped to a given semantic manifest, the valid linkable elements for a
given query should never change.