Fix distinct values rendering tests
Quick hack to get these working. Note for reviewers, this was less
silly than factoring out the common internal logic in this method
and splitting the entry-point - it's a whole lot of duplication of
mf_test_configuration and stuff no matter what, so we might as well
just jam in this conditional.