Update API for requesting dataflow plan optimization
In order to fully support predicate pushdown via the DataflowPlanOptimizer
framework we need two things:

1. Support for optimization in distinct values queries
2. The ability to share components between the DataflowPlanBuilder
and the PredicatePushdownOptimizer

This update addresses both of these concerns by doing a small restructure
of the DataflowPlanBuilder interface for accepting optimizers. Instead of
accepting a sequence of optimizer instances, the build_plan method will now
accept a sequence of optimization enumerations. Those will then be converted
to instances via the factory class added in this change.

From there the update to the distinct values plan method signature was
a trivial addition.

Note - snapshot updates should be limited to ID numbers due to the added
call to the DataflowPlanNodeOutputDataSetResolver in the distinct values
plan.