`WindowAggregationNode`: SQL rendering for cumulative metrics with non-default granularity (#1274)
Add dataflow plan node `WindowAggregationNode`. This node will be used
to re-aggregate cumulative metrics when they're queried with non-default
granularities. This PR primarily includes the SQL rendering logic and
the boilerplate code needed for new `DataflowPlanNodes`. Tests coming in
an upcoming PR.
Note that one piece of this is a placeholder waiting for the DSI change
that sets the `SqlWindowFunction` for a given cumulative metric.