Update optimizer to avoid putting window functions in group by (#1275)
When using the `WindowAggregationNode`, you will end up with a subquery
that our current `SqlRewritingSubQueryReducer` thinks can be collapsed
into the outer query. However, this would put window functions in the
group by clause, which triggers a SQL error. This PR updates the
`SqlRewritingSubQueryReducer` logic to avoid reducing in that
circumstance.