Consume new `cumulative_type_params` fields (#1293)
Completes SL-2413

Consume new `cumulative_type_params` fields (`window` and
`grain_to_date`) instead of `type_params` cumulative fields. Eventually
we will retire the `type_params` fields altogether. For now, we have a
transformation that copies the fields from `type_params` to
`cumulative_type_params`, if not yet set. This ensures that we only need
to consume the new fields here.

This is the last step needed to enable non-default granularities for
cumulative metrics in MF 🥳