Add semantic_model_origin property to LinkableElement interface (#1230)
LinkableElements have a commonality, which is that they always have
either a semantic model where they are defined, or they are virtual
constructs (either metric queries or virtual dimensions, such as
metric_time).

In order to do predicate pushdown evaluation against source nodes,
we require the linkable element to have a singular semantic model origin.
That may be a virtual model, as in the case of metric_time or metric
elements, or it may be a concrete model, as in the case of all other
entities and dimensions.

This change provides that accessor. It is separated out for ease of
review due to the need to rename the existing `semantic_model_origin`
property. We could not re-use it, as dataclasses do not allow a mixture
of @property decorated functions and simple properties. This has the
added benefit of differentiating between the semantic model origin,
which may be virtual, and the semantic model containing the element
definition.