Add PredicatePushdownOptimizer in tracking-only mode (#1262)
In order to run predicate pushdown in a more robust manner we would
be better served applying it as a post-plan-building optimizer rather
than a build-time injection of additional where constraints.

This PR adds the optimizer class with full tracking of predicate
pushdown state, but it simply returns all nodes in the input
DataflowPlan without modifying anything, since the predicate pushdown
is already applied in the DataflowPlanBuilder.

Future updates will move both the where constraint and time range constraint
pushdown operations into this optimizer class. For now we simply evaluate
the output from console logs to get a basis for future implementation changes.