Store a list of input WhereFilterSpecs in the WhereConstraintNode (#1277)
Up until now we've been merging sets of WhereFilterSpec instances
into a single instance and then storing that in the WhereConstraintNode.

This made sense for rendering, but when trying to do predicate pushdown
having the specs all merged together limits the space of filter pushdown
opporunities. In order to allow for the same breadth of predicate pushdown
opportunities we have at dataflow plan build time we keep the specs separate,
and encapsulate the merging of these specs into the WhereConstraintNode itself.