Verify column sets match in optimizer pushdown evaluation (#1279)
The PredicatePushdownOptimizer did not have access to the node
dataset resolver necessary to do the comparison between the linkable
specs referenced in the where filter and the linkable specs available
from the DataflowPlanNode targeted by predicate pushdown.

This change makes the node dataset resolver available in the optimizer.
It uses the one from the DataflowPlanBuilder in order to take advantage
of the cached resolutions available from the build process. The optimizer
is then used to evaluate the column matches in the same manner as the
original build-time pushdown evaluation.

This change was tested by running one of the predicate pushdown rendering
tests with the --log-cli-level=DEBUG confugration set, and observing the
debug output including the same entry for "Filter specs to add:" as in
the preceding commit.