Move categorical dimension filter pushdown to PredicatePushdownOptimizer (#1286)
Move categorical dimension filter pushdown to PredicatePushdownOptimizer

We are ready to move the categorical dimension filter predicate pushdown
out of the DataflowPlanBuilder and into the PredicatePushdownOptimizer.
This change makes the move with as close to parity on the existing pushdown
as possible, and adds some concrete tests of the optimizer now that it is
effecting changes on the DataflowPlan.

In theory, the outcome of this change should be that snapshots generated
from un-optimized DataflowPlans should no longer have an extra where
filter from the original build-time pushdown operation, while optimized
snapshots should be unchanged. In practice, the optimized snapshots still
have changes. These fit into two categories:

1. Subquery ID number changes caused by the removal of the extra subquery
in the non-optimized snapshots, since the ID numbers do not re-set between
the non-optimized and optimized runs.
2. Conversion metric rendering for query time filters now includes an
extra where constraint subquery due to an irrelevant pushdown operation. This is
caused by a difference in where the "disable pushdown for conversion metrics"
logic is applied - the optimizer cannot apply it until the join on conversion
node, while the original builder could effectively apply the change at compute
metric node level, and so we push down query-time filters one extra step. This
will be reverted when we remove the duplicated filters from the pushdown
operation.

This change also required fixes to a few bugs with the original tracking
implementation that only showed up when the full range of potential
snapshot updates had to be applied. In particular, a long-standing issue with
the previously never-called with_new_parents method of the CombineAggregatedOutputs
node has been resolved, and some missing or incorrectly applied checks to
prevent pushing time-based predicates down through offset windows and other
joins were addressed.