Fix bug with functionally identical comparisons on WhereConstraintNode
Due to an oversight the functionally_identical and with_new_parents
logic was not updated on the WhereConstraintNode to account for the
always_apply property.

In isolation this doesn't affect the predicate pushdown optimizer, but
in conjunction with the source scan optimizer this causes problems.

Rather than merge something broken I'm amending this PR.