Enable PredicatePushdownOptimization for all MetricFlowEngine queries
This effectively releases PredicatePushdownOptimization - the moment
this change is deployed to cloud it will be enabled.

In order to allow for a rapid mitigation of any unexpected issues this
also parameterizes the query request object to allow callers to
disable optimizations as needed. This means cloud services calling
this method can override the optimizer behaviors without requiring
an update to MetricFlow.