Improve readability of PredicatePushdownStateTracker
Added some assertions to make it more obvious what the current
branch state expectations are, and greatly expanded the documentation
of current behavior.

There is an update to the order of operations in the join handling nodes
as well, which aligns the processing with the expanded documentation in
the state tracking object.