Enable SourceScanOptimizer in query rendering tests (#1306)
Enable SourceScanOptimizer in query rendering tests

The query rendering test suite recently added support for
DataflowPlanOptimizer usage, but that only included the
PredicatePushdownOptimizer to support development.

Now that we are getting ready to release we need to see how the
queries come out with all optimizers in place, so we add the
missing SourceScanOptimizer to the set.

Update SQL engine snapshots