Upgrade DSI (#1315)
Upgrades DSI. This is needed to upgrade DSI in MFS. MFS does not depend
on DSI directly, instead relying on MF to set the DSI dependency. The
DSI upgrade is needed in MFS to get [this bug
fix](https://github.com/dbt-labs/dbt-semantic-interfaces/pull/309)
released.
[This is the PR](https://github.com/dbt-labs/metricflow-server/pull/746)
currently blocked by the upgrade.