Use `MetricTimeDefaultGranularityPattern` to resolve `metric_time` granularity (#1332)
Resolve metric time using new spec pattern in group by and filter nodes.
This is a change from defaulting to the minimum time granularity for
`metric_time`. Now we will use the max default granularity for the
requested metrics.

Note the snapshot updates. These reflect another behavior change in
which we don't error if `metric_time` is queried for metrics with two
different default granularities. Instead, we choose the larger of the
two, which is guaranteed to work for both metrics.