Updated error message for validation tasks (#1338)
## Context
Currently, in dwh validations for dimensions, we query all the
dimensions in 1 query and if that fails we attempt to query each
dimension one by one. However, with time dimension, we end up querying
multiple variants of a singular dimension (ie., `ordered_at` expands to
querying `ordered_at__week`, `ordered_at__month`, etc...), but the error
message only used the element_ment. This meant in the error, we ended up
seeing "duplicates", but it was actually just querying each one of those
variants. With this, we should use the qualfiied name to see what we're
actually querying for

#### Example:
<img width="801" alt="Screenshot 2024-07-15 at 10 33 32 PM"
src="https://github.com/user-attachments/assets/3bce8e55-49f9-43b7-9910-8076249b0eda">