Add Column Header to the exported CSV from query command (#1313)
# Issue
https://github.com/dbt-labs/metricflow/issues/1285

The CSV exported from the `mf query <metric> --csv <csv-file-name>`
command does not contain the column header. The CSV thus does not make
too much sense.

# Fix
- The code
[here](https://github.com/dbt-labs/metricflow/blob/main/dbt-metricflow/dbt_metricflow/cli/main.py#L344)
is writing data to the CSV file. The column names aren't getting added
to the file.
- With this PR, we are adding the column names returned from
`df.column_names` to the CSV to fix the issue.