Add rendering test for predicate pushdown of saved query filters
A customer reported problems with predicate pushdown removing
filters from certain subqueries. The scenario they were using
involved a saved query with multiple metrics and a set of filters
defined in the saved query itself.

This scenario represents a gap in our test coverage, so we add
a rendering test here. Note the test does not reproduce the
reported issue in any way, but it is useful to have for
the purposes of future development work so we add it here.