Bug fix: allow multiple metric input measures with the same spec (#1374)
Fixes a bug that a user encountered yesterday, which triggered our
alerts. This query was for a ratio metric `users_created /
users_created_cumulative`, a perfectly valid metric. The input measure
specs were the same, but the input measure INSTANCES had different
`cumulative_descriptions` (one is cumulative and one is not).

If interested, see the alert and DD links here:
https://dbt-labs.slack.com/archives/C05D1MF1N4Q/p1724196314740709