Add `custom_granularity_columns` to `TimeSpineSource` (#1373)
Completes SL-2744

Add `custom_granularity_columns` to `TimeSpineSource`. Along the way:
- Bumped DSI version to get the new spec that includes custom
granularities. Updated this in both MF and `metricflow-semantics`, since
the spec was needed in both places. Now that mantle points to an older
commit of MF, it should be fine to use a dev version here for now. We'll
just need to update to a non-dev version if we ever need to update the
commit that mantle is using.
- Updated the commit of `dbt-core` that we use for tests. This was
needed to enable compatibility with DSI 0.7.0.
- Renamed old attributes that weren't specific enough to differentiate
them from the new attributes.