Use ExpandedTimeGranularity in TimeDimensionSpec
In order to support custom granularities we need a way to propagate
them through the dataflow plan building and plan conversion processes.
We typically use the TimeDimensionSpec classes for carrying that
information, since the specs indicate the grain of the specific
variation of the TimeDimension element we're operating on. We
decided to use an ExpandedTimeGranularity construct to encapsulate
the custom granularity metadata in a manner interchangeable with the
default granularity enumeration values, with the idea that these
would make it more obvious when and how to do comparisons and things.

This change updates the TimeDimensionSpec itself to use the new construct.
The changes here are meant to be as mechanical as possible, so there are
places where restructuring or follow-up is needed. It was not possible to
do purely mechanical changes in all cases.