Add ExpandedTimeGranularity struct to linkable elements (#1385)
With the impending introduction of custom granularities,
which are named granularities accessed via a lookup keyed
by a date/time value with a specified base grain, our reliance
on raw TimeGranularity enumeration elements will no longer
be sufficient for managing dataflow plans or queries.

In order to be able to reference custom granularities in
our TimeDimensionSpec and associated ElementPathKey and
LinkableElement subclasses we need to make space for both
the granularity name and a base grain value.

This change adds an ExpandedTimeGranularity struct to
encapsulate both of these pieces of information in a single
object value, and switches the LinkableDimension time granularity
property to use it in place of the bare TimeGranularity enum value.

This is just the first step. Later changes will migrate most
internal TimeGranularity-typed properties to the new ExpandedTimeGranularity
construct.

Once we have a proof of concept for this working we will decide
whether this construct belongs in dbt-semantic-interfaces or if we can
keep this internal to MetricFlow. The main reason to expand this
is for more natural support for time windows and offsets, but
we may be able to simply allow strings in the protocol spec and
convert them to the struct types inside of MetricFlow.