Remove `from_name` methods on spec classes (#1386)
A number of our spec classes have a `from_name` method which
wraps the StructuredLinkableSpecName.from_name method.

In the dominant majority of cases these calls are unnecessary,
and the behavior of our code would be far more clear with a targeted
set of assertions and standard dataclass initialization.

We also have a looming issue with custom granularities affecting
the way we manage granularity configurations, which will typically
require us to initialize spec classes in a context where we have
access to the properties of any custom granularity. This can't be
readily handled from within the spec classes themselves.

This change rmeoves the `from_name` methods to avoid proliferation of
unnecessary uses, and to streamline the process of moving to direct
initialization when custom granularities are involved.