Add `JoinToCustomGranularityNode` (#1395)
Basics for adding the new `JoinToCustomGranularityNode`. We will add one
of these nodes for each custom granularity requested in the query, since
there will need to be a separate join for each one.
This PR includes the node itself and the basic visitor methods. The SQL
rendering node will be added in a follow-up commit since the logic is
more complex there. This node is never used in the dataflow plan at this
point.
@tlento I left a couple questions inline about these methods to make
sure I'm taking the right approach - I'm not completely familiar with
some of the logic for those visitors.