Add SQL rendering logic for custom granularities (#1408)
Render SQL for custom granularities.
The `JoinToCustomGranularityNode` will build a time spine dataset and
`LEFT OUTER JOIN` it to the parent dataset, without a subquery. This is
done by joining the time spine table directly, instead of creating a
subquery that selects from the time spine table and then joins. That
makes this the first node that does not create its own subquery. This
join should be safe to do without a subquery since it is always a `LEFT
OUTER JOIN`, ensuring the number of rows will not be changed by the
join, making it relatively safe for optimizers to handle.
This behavior was discussed with @tlento and agreed upon in order to
ensure optimization and readability.
You can see examples of what this SQL will look like in a [testing
PR](https://github.com/dbt-labs/metricflow/pull/1410) higher up the
stack.