SQL rendering tests for no-metric queries with custom granularity (#1411)
Tests for no-metric queries with custom granularities. Does not include
tests with filters (coming later). If you have suggestions for
additional no-metric scenarios that should be tested, please let me
know!
Note that there is a bug in the optimization for one of these tests. I
have a task to fix this up stack.