Support custom granularity in linkable spec resolution (#1415)
Further updates to linkable spec resolution needed to support custom
granularities.

This also fixes a bug related to cumulative metric restrictions:
Cumulative metrics CAN be queried with non-default granularity now, so
this removes that restriction. This was not an issue in actual queries
because that logic was only hit when querying metadata, but it could
result in a bug when listing dimensions for a cumulative metric. Instead
of restricting derived time granularity, restrict cumulative metrics
from being used with date part (this matches existing query behavior).
This change required adding a `DATE_PART` `LinkableElementProperty` and
updating `LinkableDimensions` accordingly, so that's what you'll see in
the snapshot changes.