Bug fix: custom granularity join column gets pruned (#1425)
We have an issue with the `SqlColumnPruner` where it might prune a
column that's used in the join condition, if it's not used anywhere
else. There isn't a simple way to fix that issue in the optimizer
itself, since the join condition can contain any `SqlExpressionNode`,
and this appears to only be relevant for the new custom granularity
node. Thus, I've fixed this in the custom granularity node instead of
the optimizer by rendering the column's expression in the join condition
instead of its alias.