Upgrade DSI (#1433)
Upgrade DSI in MF.
This is blocking mantle from release currently. I made a change to
dbt-core on Friday to pin core to DSI 0.7.2. This went through without
issue. But when trying to sync that change to mantle, we ran into a
version conflict with metricflow-semantics (which is only a dependency
in mantle, not core), since was pinned to a metricfllow commit SHA that
uses a lower version of DSI. Here I am upgrading the dependency in
metricflow-semantics to match DSI 0.7.2 at minimum, or any other higher
version that is non-breaking. This will allow us to continue developing
in core using new DSI versions without having to change the
metricflow-semantics commit SHA in mantle every time.
Once this commit merges, I will need to make a change to mantle to point
to this commit's SHA.

Note that the only change in the new version is adding new validations,
which won't impact MF at all.