Don't drop parent attributes when joining to custom granularity (#1439)
An oversight when building this SQL rendering logic - we were dropping
some of the SQL select node attributes from the dataset when joining to
time spine. This fixes that.
I'm not actually sure of what query scenario might have this attributes
before the custom granularity join, so this is more of a "just in case"
fix.