Cache additional methods in `MetricLookup` / `SemanticModelLookup` (#1438)
Similar to the previous PRs, this PR caches a few additional methods. I
did a little bit of memory profiling using a large manifest, and total
memory allocation did not change significantly. I believe this is
because these additional caches reference objects that are already
stored in `ValidLinkableSpecResolver` or the cached objects have low
overhead.