Reduce recursive-call overhead in `MetricTimeQueryValidationRule` (#1440)
The check in `MetricTimeQueryValidationRule` is called for every metric
in a derived metric's ancestors, so this moves expensive parts of the
check to only where it's needed and caches results when possible to
reduce runtimes. The signature for the validation rule classes were
changed, so there are a number of diff lines related to that.