Optimizer bug fix: don't prune any columns whose aliases are used in the where filter or join on condition
No tests currently trigger this bug, and there might not be any production use cases that would encounter it. But the new logic added in this PR would trigger this bug. We would apply a where filter to the time spine table, the optimizer would remove the select column, and then we would get a SQL error for the optimized query because the column used in the filter does not exist.