Add LRU cache to resolve issues with `fuctools.lru_cache` (#1446)
There is a potential memory leak when decorating class instance methods
with `fuctools.lru_cache` as `fuctools.lru_cache` holds a reference to
the class instance. This PR adds a LRU cache class that can be declared
within the instance so that it's GCed when the instance is GCed. This
will be used in later PRs to update existing caches / add new caches.